Make history-support an addon.
* Makes history-support just a standard middleware addon.
* Moves serve-files into a single file (as less logic is being done
  now).
* Removes connect-chain (it is not needed).
