+++ lib/models/project.js
@@ -66,6 +66,7 @@ Project.prototype.availableAddons = function() {
   var internalMiddlewarePath = path.join(this.root, path.relative(this.root, path.join(__dirname, '../tasks/server/middleware')));
 
   this.addIfAddon(path.join(internalMiddlewarePath, 'live-reload'), addonPackages);
+  this.addIfAddon(path.join(internalMiddlewarePath, 'history-support'), addonPackages);
   this.addIfAddon(path.join(internalMiddlewarePath, 'serve-files'), addonPackages);
   this.addIfAddon(path.join(internalMiddlewarePath, 'proxy-server'), addonPackages);
