+++ lib/tasks/server/middleware/history-support/index.js
@@ -1,12 +1,19 @@
 'use strict';
 
-// Used in serve-files middleware
-
 var path = require('path');
 var fs   = require('fs');
 
-module.exports = function(watcher) {
-  return function(req, res, next) {
+function HistorySupportAddon(project) {
+  this.project = project;
+  this.name = 'history-support-middleware';
+}
+
+HistorySupportAddon.prototype.serverMiddleware = function(config) {
+  var app = config.app;
+  var options = config.options;
+  var watcher = options.watcher;
+
+  app.use(function(req, res, next) {
     watcher.then(function(results) {
       var hasHTMLHeader = (req.headers.accept || []).indexOf('text/html') === 0;
       var isAsset = fs.existsSync(path.join(results.directory, req.path));
@@ -18,5 +25,7 @@ module.exports = function(watcher) {
 
       next();
     });
-  };
+  });
 };
+
+module.exports = HistorySupportAddon;
