+++ lib/tasks/server/middleware/serve-files/index.js
@@ -1,5 +1,7 @@
 'use strict';
 
+var cleanBaseURL            = require('../../../../utilities/clean-base-url');
+
 function ServeFilesAddon(project) {
   this.project = project;
   this.name = 'serve-files-middleware';
@@ -9,11 +11,30 @@ ServeFilesAddon.prototype.serverMiddleware = function(options) {
   var app = options.app;
   options = options.options;
 
-  var serveFilesMiddleware = require('./serve-files');
-  app.use(serveFilesMiddleware({
-    watcher: options.watcher,
-    baseURL: options.baseURL
-  }));
+  var broccoliMiddleware = options.middleware || require('broccoli/lib/middleware');
+  var middleware = broccoliMiddleware(options.watcher);
+
+  var baseURL = cleanBaseURL(options.baseURL);
+
+  app.use(function(req, res, next) {
+    var oldURL = req.url;
+
+    var actualPrefix   = req.url.slice(0, baseURL.length - 1); // Don't care
+    var expectedPrefix = baseURL.slice(0, baseURL.length - 1); // about last slash
+
+    if (actualPrefix === expectedPrefix) {
+      req.url = req.url.slice(actualPrefix.length); // Remove baseURL prefix
+
+      // Serve file, if no file has been found, reset url for proxy stuff
+      // that comes afterwards
+      middleware(req, res, function(err) {
+        req.url = oldURL;
+        next(err);
+      });
+    } else {
+      next();
+    }
+  });
 };
 
 module.exports = ServeFilesAddon;
