+++ package.json
@@ -53,7 +53,6 @@
     "chalk",
     "concat-stream",
     "configstore",
-    "connect-chain",
     "connect-livereload",
     "expresss",
     "diff",
@@ -103,7 +102,6 @@
     "chalk": "^0.4.0",
     "concat-stream": "^1.4.6",
     "configstore": "^0.2.3",
-    "connect-chain": "0.0.1",
     "connect-livereload": "^0.4.0",
     "debug": "^0.7.4",
     "diff": "^1.0.8",
